Support x86 32 bit build.  As yet untested.
Mostly fixing warnings and adding some comments.. Major difference is that extern C names are mangled under x86 so a def file is needed for the XLL and the function  register must  apply decoration to the entry point name.

+++ src/CoreStatic/ExcelArray.h
@@ -72,7 +72,7 @@ namespace xloil
     {
       if (obj.type() != ExcelType::Multi)
         XLO_THROW("Expected array");
-      if (nRows > obj.val.array.rows || nCols > obj.val.array.columns)
+      if (nRows > (size_t)obj.val.array.rows || nCols > (size_t)obj.val.array.columns)
         XLO_THROW("Out of range");
       _data = (const ExcelObj*)obj.val.array.lparray;
       _baseCols = (col_t)obj.val.array.columns;
